+++ integrations/bitbucket-cloud/integration.py
@@ -127,10 +127,10 @@ class BitbucketFilePattern(BaseModel):
 class BitbucketFileSelector(Selector):
     files: BitbucketFilePattern
     included_files: list[str] = Field(
-        title="Additional attached files (optional)",
+        title="Additional files",
         alias="includedFiles",
         default_factory=list,
-        description="List of file paths to fetch and attach to the file entity",
+        description="List of file paths to fetch and attach to the file entity. This selector will add the content of the file to the API response under the `__includedFiles` field.",,
     )